refactor: 무한스크롤 컴포넌트 분리
